requirements.txt
python-dotenv
Django
requests